Refactor agency service: split handlers into separate files and add entity types
