+++ services/agency/internal/app/entity_type.go
@@ -1,0 +1,8 @@
+package app
+
+type entityType = string
+
+const (
+	entityTypeAgency     = "AGENCY"
+	entityTypeMembership = "MEMBERSHIP"
+)
