+++ services/agency/internal/app/routes.go
@@ -11,7 +11,7 @@ import (
 func addRoutes(mux *http.ServeMux, config Config, logger *slog.Logger, client *dynamodb.Client) {
 	mux.Handle(fmt.Sprintf("GET /%s", config.Environment), listAgencies(config, logger, client))
 	mux.Handle(fmt.Sprintf("GET /%s/{id}", config.Environment), readAgency(config, logger, client))
-	mux.Handle(fmt.Sprintf("GET /%s/{id}/members", config.Environment), listAgencyMembers(config, logger, client))
+	mux.Handle(fmt.Sprintf("GET /%s/{id}/members", config.Environment), listMemberships(config, logger, client))
 
 	mux.Handle(fmt.Sprintf("POST /%s", config.Environment), createAgency(config, logger, client))
 }
