+++ services/agency/internal/app/status.go
@@ -1,0 +1,8 @@
+package app
+
+type status = string
+
+const (
+	statusActive   = "ACTIVE"
+	statusInactive = "INACTIVE"
+)
